+++ src/main/kotlin/dnd11th/blooming/common/exception/ExceptionCode.kt
@@ -1,4 +1,4 @@
-package dnd11th.blooming.common.util
+package dnd11th.blooming.common.exception
 
 enum class ExceptionCode {
     INVALID_REQUEST,
