+++ src/main/kotlin/dnd11th/blooming/common/exception/PlantNotFoundException.kt
@@ -1,6 +1,5 @@
 package dnd11th.blooming.common.exception
 
-import dnd11th.blooming.common.util.ExceptionCode
 import org.springframework.http.HttpStatus
 
 class PlantNotFoundException(
